Fix workspace container tests attempt 1
Tests pass locally but consistently fail on Github Actions. Try making
the tests themselves ensure that the services are available rather than
checking a possibly fragile log line.

Also move them to a test directory rather than scripts.

Also also slightly modify the Docker file so slow steps are cached
before copying in code, which busts the cache.
