+++ test/README.md
@@ -1,0 +1,2 @@
+This directory contains tests not part of the standard java suite. Currently it has a very small
+test suite that runs against the workspace in a docker container.
